+++ packages/shared-fixtures/fixtures/typescript/types/function-with-array-destruction.src.ts
@@ -1,0 +1,1 @@
+type foo = ([a]: any) => any