+++ packages/typescript-estree/tests/fixtures/semanticInfo/non-existent-estree-nodes.src.ts
@@ -1,0 +1,5 @@
+const binExp = (3 + 5);
+
+class Bar {
+  ['test']: string;
+}